pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "semantic-sentinel"
version = "0.0.1"
description = "Enterprise-grade Security Framework for LLMs"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  {name = "Avdhesh Singh Chouhan", email = "ai.avdhesh@gmail.com"},
]
keywords = ["security", "llm", "guardrails", "pii", "injection", "firewall", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
    "pydantic>=2.9.2",
    "pyyaml>=6.0",
    "sentence-transformers",
    "scikit-learn",
    "guardrails-ai",
    "presidio-analyzer",
    "presidio-anonymizer",
    "spacy",
    "whylogs[langkit]",
    "langkit",
    "detoxify"
]

[project.optional-dependencies]
api = [
    "fastapi>=0.115.0",
    "uvicorn>=0.31.1",
    "httpx>=0.27.2",
    "python-dotenv>=1.0.1"
]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy"
]

[project.urls]
Homepage = "https://github.com/rarenicks/secure-llm-gateway"
Documentation = "https://github.com/rarenicks/secure-llm-gateway/tree/main/docs"
Repository = "https://github.com/rarenicks/secure-llm-gateway"
Issues = "https://github.com/rarenicks/secure-llm-gateway/issues"
Changelog = "https://github.com/rarenicks/secure-llm-gateway/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["sentinel"]

# Ensure YAML profiles are included
[tool.hatch.build.targets.wheel.force-include]
"sentinel/profiles" = "sentinel/profiles"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.mypy]
strict = false
ignore_missing_imports = true
